refactor: 날짜 포맷 공통함수 추가 #3
